+++ R/apply_niche.R
@@ -54,7 +54,8 @@ apply_niche = function(x, niche_def, gc){
     #'  x = stasis_sl(seq(0, max(t), length.out = 10))
     #'  plot(reduce_to_paleoTS(x), main = "Trait evolution before niche modeling")
     #'  y = apply_niche(x, niche_def, gc)
-    #'  plot(reduce_to_paleoTS(y), main = "Trait evolution after niche modeling")
+    #'  plot(reduce_to_paleoTS(y), main = "Trait evolution after niche modeling",
+    #'  ylim = c(-2, 2))
     #'  # note that there are fewer sampling sites
     #'  # bc the taxon does not appear everywhere
     #'  # and there are fewer specimens per sampling site