+++ NEWS.md
@@ -1,3 +1,5 @@
+# StratPal (development version)
+
 # StratPal 0.6.0
 
 * Add wrappers `last_occ` and `range_offset` to determine position/timing of last occurrences & biostratigraphic precision as a function of taphonomy, ecology, abundance, and stratigraphy.